refactor(normalizer): Extract ckeditor5-core specific package.
