+++ packages/ckeditor5-common/README.md
@@ -1,5 +1,6 @@
 CKEditor 5 Common
 ================================================================================
 
-This package contains common tooling to help with CKEditor 5 Plugin
-implementations.
+This package contains generic typescript language tools independent to CKEditor 5.
+
+It is very important that this package does not have any dependency to ckeditor 5 packages.
