+++ packages/ckeditor5-core-common/README.md
@@ -1,0 +1,5 @@
+CKEditor 5 Core Common
+================================================================================
+
+This package contains common tooling to help with CKEditor 5 Plugin
+implementations.
