+++ packages/ckeditor5-core-common/babel.config.js
@@ -1,0 +1,3 @@
+const sharedBabelConf = require('@coremedia-internal/ckeditor5-babel-config');
+
+module.exports = sharedBabelConf;
