+++ packages/ckeditor5-core-common/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@coremedia/ckeditor5-common",
+  "name": "@coremedia/ckeditor5-core-common",
   "version": "6.1.1-rc.1",
   "description": "tooling to help with CKEditor 5 Plugin implementations.",
   "keywords": [
@@ -16,6 +16,7 @@
   "devDependencies": {
     "@babel/cli": "^7.17.10",
     "@babel/core": "^7.18.2",
+    "@ckeditor/ckeditor5-core": "^34.1.0",
     "@coremedia-internal/ckeditor5-babel-config": "^1.0.0",
     "@coremedia-internal/ckeditor5-jest-test-helpers": "^1.0.0",
     "@types/ckeditor__ckeditor5-core": "^33.0.3",
@@ -32,8 +33,12 @@
     "rimraf": "^3.0.2",
     "typescript": "^4.7.2"
   },
+  "peerDependencies": {
+    "@ckeditor/ckeditor5-core": "^34.1.0"
+  },
   "dependencies": {
-    "@coremedia/ckeditor5-logging": "6.1.1-rc.1"
+    "@coremedia/ckeditor5-logging": "6.1.1-rc.1",
+    "@coremedia/ckeditor5-common": "6.1.1-rc.1"
   },
   "scripts": {
     "clean": "rimraf ./dist",
