+++ packages/ckeditor5-core-common/src/Commands.ts
