+++ packages/ckeditor5-core-common/src/Plugins.ts
