+++ packages/ckeditor5-core-common/src/index-doc.ts
@@ -1,0 +1,7 @@
+/**
+ * Tooling to help with CKEditor 5 Plugin implementations.
+ * @module ckeditor5-core-common
+ */
+export * from "./Commands";
+export * from "./Editors";
+export * from "./Plugins";
