+++ packages/ckeditor5-core-common/src/tsconfig.json
@@ -1,0 +1,23 @@
+{
+  "compilerOptions": {
+    "baseUrl": ".",
+    "rootDir": ".",
+    "outDir": "../dist",
+    "emitDeclarationOnly": true,
+    "declaration": true,
+    "declarationMap": true,
+    "paths": {
+      "@coremedia/ckeditor5-logging/*": ["../../ckeditor5-logging/src/*"],
+      "@coremedia/ckeditor5-common/*": ["../../ckeditor5-common/src/*"]
+    }
+  },
+  "references": [
+    {
+      "path": "../../ckeditor5-logging/src"
+    },
+    {
+      "path": "../../ckeditor5-common/src"
+    }
+  ],
+  "extends": "../../../tsconfig.json"
+}
