+++ packages/ckeditor5-coremedia-images/package.json
@@ -52,7 +52,7 @@
     "@coremedia/service-agent": "^1.1.0"
   },
   "dependencies": {
-    "@coremedia/ckeditor5-common": "6.1.1-rc.1",
+    "@coremedia/ckeditor5-core-common": "6.1.1-rc.1",
     "@coremedia/ckeditor5-coremedia-content-clipboard": "6.1.1-rc.1",
     "@coremedia/ckeditor5-coremedia-studio-integration": "6.1.1-rc.1",
     "@coremedia/ckeditor5-logging": "6.1.1-rc.1",
