+++ packages/ckeditor5-coremedia-images/src/ContentImageClipboardPlugin.ts
@@ -11,7 +11,7 @@ import {
   CreateModelFunctionCreator,
 } from "@coremedia/ckeditor5-coremedia-content-clipboard/ContentToModelRegistry";
 import ContentClipboardEditing from "@coremedia/ckeditor5-coremedia-content-clipboard/ContentClipboardEditing";
-import { ifPlugin, recommendPlugin } from "@coremedia/ckeditor5-common/Plugins";
+import { ifPlugin, recommendPlugin } from "@coremedia/ckeditor5-core-common/Plugins";
 
 type CreateImageModelFunction = (blobUriPath: string) => CreateModelFunction;
