+++ packages/ckeditor5-coremedia-link/package.json
@@ -63,6 +63,7 @@
   },
   "dependencies": {
     "@coremedia/ckeditor5-common": "6.1.1-rc.1",
+    "@coremedia/ckeditor5-core-common": "6.1.1-rc.1",
     "@coremedia/ckeditor5-coremedia-content-clipboard": "6.1.1-rc.1",
     "@coremedia/ckeditor5-coremedia-studio-integration": "6.1.1-rc.1",
     "@coremedia/ckeditor5-logging": "6.1.1-rc.1",
