+++ packages/ckeditor5-coremedia-link/src/tsconfig.json
@@ -8,6 +8,7 @@
     "declarationMap": true,
     "paths": {
       "@coremedia/ckeditor5-common/*": ["../../ckeditor5-common/src/*"],
+      "@coremedia/ckeditor5-core-common/*": ["../../ckeditor5-core-common/src/*"],
       "@coremedia/ckeditor5-logging/*": ["../../ckeditor5-logging/src/*"],
       "@coremedia/ckeditor5-coremedia-content-clipboard/*": ["../../ckeditor5-coremedia-content-clipboard/src/*"],
       "@coremedia/ckeditor5-coremedia-studio-integration/*": ["../../ckeditor5-coremedia-studio-integration/src/*"]
@@ -16,6 +17,9 @@
   "references": [
     {
       "path": "../../ckeditor5-common/src"
+    },
+    {
+      "path": "../../ckeditor5-core-common/src"
     },
     {
       "path": "../../ckeditor5-logging/src"
