+++ typedoc-tsconfig.json
@@ -4,6 +4,9 @@
       "@coremedia/ckeditor5-common/*": [
         "./packages/ckeditor5-common/src/*"
       ],
+      "@coremedia/ckeditor5-core-common/*": [
+        "./packages/ckeditor5-core-common/src/*"
+      ],
       "@coremedia/ckeditor5-coremedia-content-clipboard/*": [
         "./packages/ckeditor5-coremedia-content-clipboard/src/*"
       ],
